Distributed the code according to the corresponding files #7
