+++ components/selectfilter/index.js
@@ -1,0 +1,3 @@
+module.exports = require('./component')
+module.exports.actions = require('./actions')
+module.exports.reducer = require('./reducer')
